+++ packages/core/votes/slottracker/epochtracker_test.go
@@ -1,4 +1,4 @@
-package epochtracker
+package slottracker
 
 import (
 	"testing"
