+++ packages/core/votes/slottracker/testframework.go
@@ -1,4 +1,4 @@
-package epochtracker
+package slottracker
 
 import (
 	"testing"
