+++ plugins/webapi.go
@@ -6,7 +6,6 @@ import (
 	"github.com/iotaledger/goshimmer/plugins/webapi/autopeering"
 	"github.com/iotaledger/goshimmer/plugins/webapi/block"
 	"github.com/iotaledger/goshimmer/plugins/webapi/data"
-	"github.com/iotaledger/goshimmer/plugins/webapi/epoch"
 	"github.com/iotaledger/goshimmer/plugins/webapi/faucet"
 	"github.com/iotaledger/goshimmer/plugins/webapi/faucetrequest"
 	"github.com/iotaledger/goshimmer/plugins/webapi/healthz"
@@ -15,6 +14,7 @@ import (
 	"github.com/iotaledger/goshimmer/plugins/webapi/mana"
 	"github.com/iotaledger/goshimmer/plugins/webapi/ratesetter"
 	"github.com/iotaledger/goshimmer/plugins/webapi/scheduler"
+	"github.com/iotaledger/goshimmer/plugins/webapi/slot"
 	"github.com/iotaledger/goshimmer/plugins/webapi/weightprovider"
 )
 
@@ -28,7 +28,7 @@ var WebAPI = node.Plugins(
 	block.Plugin,
 	autopeering.Plugin,
 	info.Plugin,
-	epoch.Plugin,
+	slot.Plugin,
 	mana.Plugin,
 	ledgerstate.Plugin,
 	// snapshot.Plugin,
